[stand_cloud] Mv extra func to utils
